core_modules/AccessManagementService/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==1.10.13
pyyaml>=6.0
PyJWT>=2.8.0
requests>=2.31.0
